requirements.txt
protobuf
grpcio
googleapis-common-protos